+++ Encoding/x265Enc.vb
@@ -668,7 +668,7 @@ Public Class x265Params
         .Text = "Custom",
         .AlwaysOn = True,
         .InitAction = Sub(tb)
-                          tb.Edit.Expandet = True
+                          tb.Edit.Expand = True
                           tb.Edit.TextBox.Multiline = True
                           tb.Edit.MultilineHeightFactor = 6
                           tb.Edit.TextBox.Font = New Font("Consolas", 10 * s.UIScaleFactor)
@@ -677,7 +677,7 @@ Public Class x265Params
     Property CustomFirstPass As New StringParam With {
         .Text = "Custom" + BR + "First Pass",
         .InitAction = Sub(tb)
-                          tb.Edit.Expandet = True
+                          tb.Edit.Expand = True
                           tb.Edit.TextBox.Multiline = True
                           tb.Edit.MultilineHeightFactor = 6
                           tb.Edit.TextBox.Font = New Font("Consolas", 10 * s.UIScaleFactor)
@@ -686,7 +686,7 @@ Public Class x265Params
     Property CustomSecondPass As New StringParam With {
         .Text = "Custom" + BR + "Second Pass",
         .InitAction = Sub(tb)
-                          tb.Edit.Expandet = True
+                          tb.Edit.Expand = True
                           tb.Edit.TextBox.Multiline = True
                           tb.Edit.MultilineHeightFactor = 6
                           tb.Edit.TextBox.Font = New Font("Consolas", 10 * s.UIScaleFactor)