+++ Forms/PreviewForm.vb
@@ -327,20 +327,14 @@ Public Class PreviewForm
         AVI = New AVIFile(AviSynthDocument.Path)
         Drawer = New VideoDrawer(pVideo, AVI)
         Drawer.ShowInfos = s.PreviewToggleInfos
-
         Dim wa = Screen.FromControl(Me).WorkingArea
 
-        While GetNormalSize.Width < wa.Width * 0.6
+        While GetNormalSize.Width < wa.Width * (s.MinPreviewSize / 100) AndAlso GetNormalSize.Height < wa.Height * (s.MinPreviewSize / 100)
             SizeFactor += 0.1
             SizeFactor = Math.Round(SizeFactor, 2)
         End While
 
-        While GetNormalSize.Width > wa.Width * 0.9
-            SizeFactor -= 0.1
-            SizeFactor = Math.Round(SizeFactor, 2)
-        End While
-
-        While GetNormalSize.Height > wa.Height * 0.9
+        While GetNormalSize.Width > wa.Width * 0.9 OrElse GetNormalSize.Height > wa.Height * 0.9
             SizeFactor -= 0.1
             SizeFactor = Math.Round(SizeFactor, 2)
         End While
@@ -350,6 +344,10 @@ Public Class PreviewForm
         Else
             NormalScreen()
         End If
+
+        If Left + Width > wa.Width OrElse Top + Height > wa.Height Then WindowPositions.CenterScreen(Me)
+        If Left < 0 Then Left = 0
+        If Top < 0 Then Top = 0
 
         If s.LastPosition < AVI.FrameCount - 1 Then AVI.Position = s.LastPosition
         AfterPositionChanged()
@@ -622,7 +620,9 @@ Public Class PreviewForm
     Private Sub AfterPositionChanged()
         s.LastPosition = AVI.Position
         DrawTrack()
-        Text = "Preview " & s.LastPosition
+        Dim time = TimeSpan.FromSeconds(AVI.Position / AVI.FrameRate).ToString.TrimEnd("0"c)
+        If time.StartsWith("00") Then time = time.Substring(3)
+        Text = "Preview  " & s.LastPosition & "  " + time
     End Sub
 
     <Command("Dialog to jump to a specific time.")>
@@ -1057,5 +1057,12 @@ Public Class PreviewForm
         e.Cancel = True
         MyBase.OnHelpButtonClicked(e)
         OpenHelp()
+    End Sub
+
+    Private Sub pVideo_MouseDown(sender As Object, e As MouseEventArgs) Handles pVideo.MouseDown
+        Dim sb = Screen.FromControl(Me).Bounds
+        Dim p1 = New Point(sb.Width, 0)
+        Dim p2 = PointToScreen(e.Location)
+        If Math.Abs(p1.X - p2.X) < 10 AndAlso Math.Abs(p1.Y - p2.Y) < 10 Then Close()
     End Sub
 End Class